requirements.txt
django>=2.0,<2.1
django-model-utils>=2.0
arrow>=0.12
# Additional requirements go here
